Extract capabilities, attributes and commands in own file
